third_party/blink/web_tests/accessibility/is-richly-editable-expected.txt

This tests that only elements that allow the user to edit rich text in them are recognized as being richly editable.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axElement.isRichlyEditable is true
PASS axElement.isRichlyEditable is true
PASS axElement.isRichlyEditable is true
PASS axElement.isRichlyEditable is true
PASS axElement.isRichlyEditable is true
PASS axElement.isRichlyEditable is true
PASS axElement.isRichlyEditable is true
PASS axElement.isRichlyEditable is false
PASS axElement.isRichlyEditable is false
PASS axElement.isRichlyEditable is false
PASS axElement.isRichlyEditable is false
PASS axElement.isRichlyEditable is false
PASS axElement.isRichlyEditable is false
PASS axElement.isRichlyEditable is false
PASS successfullyParsed is true

TEST COMPLETE


Some text.

Some text.

Some text.
 

    

Some text.




Some text.

Button
